+++ contracts/test/GatedMint.spec.ts
@@ -1,17 +1,16 @@
 import { expect } from "chai";
-import hre, { deployments } from "hardhat";
+import hre, { deployments, ethers } from "hardhat";
 import "@nomiclabs/hardhat-ethers";
 
 import { MerkleTree } from "merkletreejs";
 import { keccak256 } from "ethers/lib/utils";
-
-import { getWildcardPermit, getPermitSigners } from "./createMintPermit";
+import assert from "assert";
 
 describe("GatedMint", async () => {
   const baseSetup = deployments.createFixture(async () => {
     await deployments.fixture();
 
-    const permitSigners = getPermitSigners();
+    const permitSigners = getIssuers();
 
     const elements = await Promise.all(
       permitSigners.map((s) => s.getAddress())
@@ -28,16 +27,16 @@ describe("GatedMint", async () => {
     return { gatedMint };
   });
 
-  describe("preMint", function () {
+  describe("redeem", function () {
     it("it executes for a valid permit", async function () {
       const { gatedMint } = await baseSetup();
 
-      const signerIndex = 0;
+      const issuerIndex = 0;
       const minterIndex = 0;
 
       const signer = (await hre.ethers.getSigners())[minterIndex];
 
-      const permit = await getWildcardPermit(signerIndex, minterIndex);
+      const permit = await getWildcardPermit(issuerIndex, minterIndex);
 
       await expect(gatedMint.connect(signer)._redeem(permit)).to.not.be
         .reverted;
@@ -46,13 +45,13 @@ describe("GatedMint", async () => {
     it("reverts for a permit produced for a different minter", async function () {
       const { gatedMint } = await baseSetup();
 
-      const signerIndex = 0;
+      const issuerIndex = 0;
       const minterIndex = 0;
       const wrongMinterIndex = 1;
 
       const signer = (await hre.ethers.getSigners())[wrongMinterIndex];
 
-      const permit = await getWildcardPermit(signerIndex, minterIndex);
+      const permit = await getWildcardPermit(issuerIndex, minterIndex);
 
       await expect(
         gatedMint.connect(signer)._redeem(permit)
@@ -62,12 +61,11 @@ describe("GatedMint", async () => {
     it("reverts for an already used proof", async function () {
       const { gatedMint } = await baseSetup();
 
-      const signerIndex = 0;
+      const issuerIndex = 0;
       const minterIndex = 0;
 
       const signer = (await hre.ethers.getSigners())[minterIndex];
-
-      const permit = await getWildcardPermit(signerIndex, minterIndex);
+      const permit = await getWildcardPermit(issuerIndex, minterIndex);
 
       await expect(gatedMint.connect(signer)._redeem(permit)).to.not.be
         .reverted;
@@ -78,3 +76,40 @@ describe("GatedMint", async () => {
     });
   });
 });
+
+function getIssuers() {
+  const passwords = [
+    "double custom typical style harvest buzz",
+    "wasp decorate badge coast park pole",
+    "sense pudding cause leave security subway",
+  ];
+
+  return passwords
+    .map((password) =>
+      ethers.utils.keccak256(ethers.utils.toUtf8Bytes(password))
+    )
+    .map((pk) => new ethers.Wallet(pk));
+}
+
+async function getWildcardPermit(from: number, to: number) {
+  const elements = getIssuers().map((s) => s.address);
+  const merkleTree = new MerkleTree(elements, keccak256, {
+    hashLeaves: true,
+    sortPairs: true,
+  });
+
+  const issuer = getIssuers()[from];
+
+  const minter = (await hre.ethers.getSigners())[to];
+
+  // not we have to arrayify because wallet.signMessage does not recognize hex strings
+  const messageHash = ethers.utils.arrayify(keccak256(minter.address));
+  const signature = await issuer.signMessage(messageHash);
+
+  assert(ethers.utils.verifyMessage(messageHash, signature) == issuer.address);
+
+  return {
+    signature,
+    proof: merkleTree.getHexProof(keccak256(issuer.address)),
+  };
+}
